Run SettingMacros as a manual filter before markdown
The markdown filter will break on creating links from the new syntax
