+++ lib/open_project/text_formatting/filters/setting_macros_filter.rb
@@ -30,23 +30,47 @@
 
 module OpenProject::TextFormatting
   module Filters
-    class SettingMacrosFilter < PatternMatcherFilter
-      def self.matchers
-        [
-          OpenProject::TextFormatting::Matchers::SettingMacros
-        ]
+    class SettingMacrosFilter < HTML::Pipeline::Filter
+      ALLOWED_SETTINGS = %w[
+        host_name
+        base_url
+      ].freeze
+
+      def self.regexp
+        %r{
+        \{\{opSetting:(.+?)\}\}
+        }x
       end
 
       def call
-        doc.search('.//text()|.//a[@href]').each do |node|
-          next if has_ancestor?(node, PREFORMATTED_BLOCKS)
+        return html unless applicable?
 
-          self.class.matchers.each do |matcher|
-            matcher.call(node, doc: doc, context: context)
+        html.gsub(self.class.regexp) do |matched_string|
+          variable = $1.to_s
+          variable.gsub!('\\', '')
+
+          if ALLOWED_SETTINGS.include?(variable)
+            send variable
+          else
+            matched_string
           end
         end
+      end
 
-        doc
+      private
+
+      def host_name
+        OpenProject::StaticRouting::UrlHelpers.host
+      end
+
+      def base_url
+        OpenProject::Application.root_url
+      end
+
+      ##
+      # Faster inclusion check before the regex is being applied
+      def applicable?
+        html.include?('{{opSetting:')
       end
     end
   end
